All tests pass! Removed unused code. Renamed package.
