+++ core/src/main/java/org/neo4j/ogm/metadata/builder/AnnotationsInfoBuilder.java
@@ -1,4 +1,4 @@
-package org.neo4j.ogm.metadata.reflections;
+package org.neo4j.ogm.metadata.builder;
 
 import java.lang.annotation.Annotation;
 import java.util.HashMap;
