+++ core/src/main/java/org/neo4j/ogm/metadata/builder/InterfacesInfoBuilder.java
@@ -1,4 +1,4 @@
-package org.neo4j.ogm.metadata.reflections;
+package org.neo4j.ogm.metadata.builder;
 
 import java.util.HashMap;
 import java.util.Map;
