--- core/src/main/java/org/neo4j/ogm/metadata/bytecode/ResourceResolver.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2002-2016 "Neo Technology,"
- * Network Engine for Objects in Lund AB [http://neotechnology.com]
- *
- * This product is licensed to you under the Apache License, Version 2.0 (the "License").
- * You may not use this product except in compliance with the License.
- *
- * This product may include a number of subcomponents with
- * separate copyright notices and license terms. Your use of the source
- * code for these subcomponents is subject to the terms and
- *  conditions of the subcomponent's license, as noted in the LICENSE file.
- */
-
-package org.neo4j.ogm.metadata.bytecode;
-
-import java.io.File;
-import java.net.URL;
-
-/**
- * @author vince
- */
-public interface ResourceResolver
-{
-    /**
-     * If a ResourceResolver understands the protocol of a specified resource
-     * it should resolve the resource as a File, and return it.
-     *
-     * If it is unable to convert the resource, it should return null.
-     *
-     * @param resource the URL to be converted
-     * @return the resource as a File
-     */
-    File resolve( URL resource ) throws Exception;
-}
