+++ core/src/test/java/org/neo4j/ogm/metadata/DomainInfoTest.java
@@ -14,7 +14,7 @@ package org.neo4j.ogm.metadata;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.neo4j.ogm.metadata.reflections.DomainInfoBuilder;
+import org.neo4j.ogm.metadata.builder.DomainInfoBuilder;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
