+++ core/src/test/java/org/neo4j/ogm/persistence/relationships/transitive/ab/ABTest.java
@@ -44,7 +44,7 @@ public class ABTest extends MultiDriverTestClass {
 
     @BeforeClass
     public static void oneTimeSetUp() {
-        sessionFactory = new SessionFactory("org.neo4j.ogm.persistence.relationships.transitive.ab");
+        sessionFactory = new SessionFactory("org.neo4j.ogm.persistence.relationships.transitive.ab.");
     }
 
     @Before
